+++ src/base/object/index.ts
@@ -1,5 +1,6 @@
 export { getFlattenedDeps } from './collect-deps';
 export { compareTwoObj } from './compare';
+export { deepSet } from './deep-set';
 export { isObjectEqual } from './equal';
 export { extend, toHumpObj } from './object';
 export { traverseResp } from './parse-interface-data';