+++ src/component-demo/user.ts
@@ -1,0 +1,6 @@
+export function getUserComponentDemo() {
+  return {
+    avatar: 'https://thirdqq.qlogo.cn/g?b=oidb&k=HFh953g22kPgQny1h7KITQ&kti=ZJJzkQAAAAE&s=640&t=1556998035',
+    nick: '杨',
+  };
+}