+++ src/lodash-mini/index.ts
@@ -1,3 +1,4 @@
 export { get } from './get';
 export { merge } from './merge';
+export { setWithString } from './set-with-string';
 export { set } from './set';