+++ src/lodash-mini/set.ts
@@ -112,3 +112,5 @@ export function merge<T extends object>(...sources: T[]) {
 
   return clone;
 }
+
+