+++ src/stylelint-rules/index.ts
@@ -1,0 +1,1 @@
+export { getStylelintVendorPrefixRules } from './vendor-prefix';