+++ src/stylelint-rules/vendor-prefix.ts
@@ -1,0 +1,9 @@
+export function getStylelintVendorPrefixRules() {
+  return {
+    'selector-no-vendor-prefix': null,
+    'value-no-vendor-prefix': null,
+    'property-no-vendor-prefix': null,
+    'at-rule-no-vendor-prefix': null,
+    'media-feature-name-no-vendor-prefix': null,
+  };
+}