+++ src/validate/index.ts
@@ -1,3 +1,4 @@
 export * from './id-card';
+export * from './is-index';
 export * from './type';
 export * from './validate';